+++ src/Traits/HasContext.php
@@ -18,6 +18,15 @@ trait HasContext
         return $this;
     }
 
+    /** @psalm-suppress PossiblyUnusedReturnValue */
+    #[\Override]
+    public function unsetContext(string $key): static
+    {
+        unset($this->_context[$key]);
+
+        return $this;
+    }
+
     public function _withContext(array $values): static
     {
         foreach ($values as $key => $val) {
@@ -31,6 +40,19 @@ trait HasContext
     public function getContext(string $key, mixed $default = null): mixed
     {
         return $this->_context[$key] ?? $default;
+    }
+
+    public function hasContext(string $key, bool $treatNullAsMissing = true): bool
+    {
+        if (!array_key_exists($key, $this->_context)) {
+            return false;
+        }
+
+        if ($treatNullAsMissing && null === $this->_context[$key]) {
+            return false;
+        }
+
+        return true;
     }
 
     public function getContextMap(): array